+++ app/Http/Controllers/FavoriteController.php
@@ -2,11 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-
 class FavoriteController extends Controller
 {
-    public function store(Request $request, $id)
+    public function store($id)
     {
         \Auth::user()->favorite($id);
         return back();
